% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tier_protocol}
\alias{tier_protocol}
\title{tier_protocol}
\usage{
tier_protocol(tier = TRUE)
}
\arguments{
\item{tier}{Defaults to TRUE.}
}
\value{
when TRUE, returns the 4.0 version of TIER protocol.

When false, returns a minimal version of TIER protocol.
}
\description{
Provides documentation for your R project,
through version 4.0 of the TIER protocol, developed by project TIER.
}
